Fixing typeo and using size_t to compare against a size() method.

+++ rai/node/node.hpp
@@ -106,7 +106,7 @@ public:
 	// After this many successive vote announcements, block is confirmed
 	static unsigned constexpr contiguous_announcements = 4;
 	static unsigned constexpr announce_interval_ms = (rai::rai_network == rai::rai_networks::rai_test_network) ? 10 : 16000;
-	static unsigned constexpr election_history_size = 2048;
+	static size_t constexpr election_history_size = 2048;
 };
 class operation
 {
